+++ app/_api/api.ts
@@ -1,7 +1,9 @@
 import { NewsClusters } from "../_types";
+import { PressBiasRequest } from "../_types/press_bias_request";
 
 const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
-async function fetchNewsClusters(): Promise<NewsClusters> {
+
+export async function getNewsClusters(): Promise<NewsClusters> {
   const response = await fetch(`${BASE_URL}/news`, {
     cache: "no-store",
   });
@@ -9,9 +11,31 @@ async function fetchNewsClusters(): Promise<NewsClusters> {
   if (!response.ok) {
     throw new Error(`HTTP error! Status: ${response.status}`);
   }
-  console.log(response);
+
   const data: NewsClusters = await response.json(); // JSON 데이터를 변환
   return data;
 }
 
-export default fetchNewsClusters;
+export async function postBiasRequest(requestBody: PressBiasRequest) {
+  try {
+    const response = await fetch(
+      `${process.env.NEXT_PUBLIC_BASE_URL}/press-bias-request`,
+      {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(requestBody),
+      }
+    );
+
+    if (response.ok) {
+      return response;
+    } else {
+      alert("요청 제출에 실패했습니다. 다시 시도해주세요.");
+    }
+  } catch (error) {
+    console.error("Error submitting request:", error);
+    alert("요청 제출 중 오류가 발생했습니다.");
+  }
+}
