+++ app/_components/NewsCard/NewsCard.module.css
@@ -1,0 +1,470 @@
+.newsCard {
+  background-color: white;
+  border-radius: 10px;
+  overflow: hidden;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.05);
+  transition: transform 0.3s, box-shadow 0.3s;
+}
+
+.newsCard:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
+}
+
+.newsImage {
+  position: relative;
+  border-top-right-radius: 8px;
+  border-top-left-radius: 8px;
+  overflow: hidden;
+}
+.imageContainer {
+  width: 100%;
+  position: relative;
+}
+
+.imageSkeleton {
+  width: 100%;
+  height: 400px;
+  background: linear-gradient(90deg, #f0f0f0 25%, #e0e0e0 50%, #f0f0f0 75%);
+  background-size: 200% 100%;
+  animation: shimmer 1.5s infinite;
+  border-top-right-radius: 8px;
+  border-top-left-radius: 8px;
+}
+
+@keyframes shimmer {
+  0% {
+    background-position: 200% 0;
+  }
+  100% {
+    background-position: -200% 0;
+  }
+}
+.hidden {
+  display: none;
+}
+
+.fullWidthImage {
+  width: 100%;
+  height: auto;
+  max-height: 600px;
+  min-height: 300px;
+  display: block;
+  object-fit: cover;
+}
+
+.cardShadow {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  height: 100px;
+  background: linear-gradient(to top, rgba(0, 0, 0, 0.6), transparent);
+  pointer-events: none;
+}
+
+.headline {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  padding: 15px;
+  color: white;
+  background: linear-gradient(
+    to bottom,
+    transparent 0%,
+    rgba(0, 0, 0, 0.5) 30%,
+    rgba(0, 0, 0, 0.7) 100%
+  );
+}
+
+.headline h2 {
+  margin: 0 0 10px 0;
+  font-size: 1.2rem;
+  line-height: 1.4;
+}
+
+.headlineMeta {
+  display: flex;
+  justify-content: space-between;
+  font-size: 0.8rem;
+  color: rgb(219, 219, 219);
+}
+.keyword {
+  font-size: 0.8rem;
+  color: rgb(219, 219, 219);
+}
+
+.coverageBar {
+  height: 50px;
+  display: flex;
+  margin-bottom: 1px;
+  width: 100%;
+}
+
+.headerContent {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding-top: 10px;
+}
+
+.logo {
+  font-size: 28px;
+  font-weight: 700;
+  color: var(--primary-color);
+}
+
+.logo span {
+  color: #444;
+}
+
+.searchBar {
+  margin: 30px 0;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.searchInput {
+  display: flex;
+  flex-grow: 1;
+  margin-right: 20px;
+  position: relative;
+}
+
+.searchInput input {
+  width: 100%;
+  padding: 12px 20px;
+  border: 1px solid var(--border-color);
+  border-radius: 30px;
+  font-size: 16px;
+  outline: none;
+  transition: box-shadow 0.3s;
+}
+
+.searchInput input:focus {
+  box-shadow: 0 0 0 2px rgba(51, 102, 204, 0.2);
+}
+
+.searchInput button {
+  position: absolute;
+  right: 5px;
+  top: 5px;
+  background-color: var(--primary-color);
+  color: white;
+  border: none;
+  border-radius: 30px;
+  padding: 8px 20px;
+  cursor: pointer;
+  font-weight: 500;
+}
+
+.filters {
+  display: flex;
+  gap: 10px;
+}
+
+.filterButton {
+  padding: 8px 15px;
+  background-color: white;
+  border: 1px solid var(--border-color);
+  border-radius: 20px;
+  cursor: pointer;
+  transition: all 0.3s;
+}
+
+.filterButton:hover {
+  background-color: var(--light-gray);
+}
+
+.filterButtonActive {
+  background-color: var(--primary-color);
+  color: white;
+  border-color: var(--primary-color);
+}
+
+/* .headline {
+  padding: 20px;
+  border-bottom: 1px solid var(--border-color);
+}
+
+.headline h2 {
+  font-size: 24px;
+  margin-bottom: 10px;
+}
+
+.headlineMeta {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  color: #666;
+  font-size: 14px;
+} */
+
+.coverageLiberal,
+.coverageNeutral,
+.coverageConservative {
+  color: white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-weight: 500;
+  transition: all 0.3s ease;
+}
+
+.coverageLiberal {
+  background-color: var(--liberal-color);
+}
+
+.coverageConservative {
+  background-color: var(--conservative-color);
+}
+
+.coverageNeutral {
+  background-color: var(--neutral-color);
+}
+
+@media (max-width: 768px) {
+  .coverageBar {
+    height: 40px;
+  }
+
+  .coverageLiberal,
+  .coverageNeutral,
+  .coverageConservative {
+    font-size: 14px;
+  }
+
+  .coverageLiberal span,
+  .coverageNeutral span,
+  .coverageConservative span {
+    display: none;
+  }
+
+  .coverageLiberal::after,
+  .coverageNeutral::after,
+  .coverageConservative::after {
+    content: attr(data-percent);
+  }
+}
+
+.articlesContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  padding: 20px;
+  width: 100%;
+}
+
+.representativeArticles {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 20px;
+  width: 100%;
+}
+
+.article {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+  padding: 20px;
+  background-color: white;
+  border: 1px solid var(--border-color);
+  border-radius: 8px;
+  width: 100%;
+}
+
+.articleHeader {
+  display: flex;
+  align-items: center;
+  margin-bottom: 15px;
+}
+
+.articleSourceImg {
+  width: 40px;
+  height: 40px;
+  border-radius: 5px;
+  margin-right: 10px;
+  object-fit: cover;
+}
+
+.articleSourceInfo h3 {
+  font-size: 16px;
+  margin: 0;
+}
+
+.articleSourceInfo .sourceBias {
+  font-size: 14px;
+  padding: 2px 8px;
+  border-radius: 10px;
+  display: inline-block;
+  margin-top: 5px;
+}
+
+.biasLiberal {
+  background-color: rgba(52, 152, 219, 0.2);
+  color: var(--liberal-color);
+}
+
+.biasConservative {
+  background-color: rgba(231, 76, 60, 0.2);
+  color: var(--conservative-color);
+}
+
+.biasNeutral {
+  background-color: var(--neutral-color);
+}
+
+.articleTitle {
+  font-size: 18px;
+  margin-bottom: 10px;
+  line-height: 1.4;
+}
+
+.articleExcerpt {
+  font-size: 14px;
+  color: #666;
+  margin-bottom: 15px;
+}
+
+.readMore {
+  color: var(--primary-color);
+  text-decoration: none;
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.footer {
+  background-color: #333;
+  color: white;
+  padding: 40px 0;
+}
+
+.footerContent {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  gap: 30px;
+  margin-top: 40px;
+  margin-bottom: 20px;
+}
+
+.footerColumn h3 {
+  font-size: 18px;
+  margin-bottom: 20px;
+}
+
+.footerLinks {
+  list-style: none;
+}
+
+.footerLinks li {
+  margin-bottom: 10px;
+}
+
+.footerLinks a {
+  color: #ccc;
+  text-decoration: none;
+  transition: color 0.3s;
+}
+
+.footerLinks a:hover {
+  color: white;
+}
+
+.copyright {
+  text-align: center;
+  margin-top: 40px;
+  color: #999;
+  font-size: 14px;
+}
+
+.showMoreButton,
+.showLessButton {
+  width: 100%;
+  padding: 12px;
+  margin: 10px 0;
+  background-color: var(--light-gray);
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  font-weight: 500;
+  transition: background-color 0.3s;
+}
+
+.showMoreButton:hover,
+.showLessButton:hover {
+  background-color: #e6e6e6;
+}
+
+.allArticles {
+  margin-top: 20px;
+  border-top: 1px solid var(--border-color);
+  padding-top: 20px;
+}
+
+.allArticles .article {
+  border: 1px solid var(--border-color);
+  margin-bottom: 15px;
+}
+
+.allArticles .article:last-child {
+  margin-bottom: 0;
+}
+
+.footerLink {
+  color: #666;
+  text-decoration: none;
+}
+
+.footerLink:hover {
+  color: #333;
+}
+
+/* .notFound {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 400px;
+  width: 100vw;
+} */
+.notFound {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: calc(100vh - 200px); /* header와 footer 높이를 고려 */
+}
+
+.notFoundContent {
+  text-align: center;
+  padding: 2rem;
+}
+
+.notFoundIcon {
+  color: var(--primary-color, #336699);
+  margin-bottom: 2rem;
+  animation: float 3s ease-in-out infinite;
+}
+
+.notFoundContent h1 {
+  font-size: 2rem;
+  margin-bottom: 1rem;
+  color: #333;
+}
+
+.notFoundContent p {
+  color: #666;
+  font-size: 1.1rem;
+}
+
+@keyframes float {
+  0% {
+    transform: translateY(0px);
+  }
+  50% {
+    transform: translateY(-10px);
+  }
+  100% {
+    transform: translateY(0px);
+  }
+}
