+++ app/_components/TopKeywordContainer/TopKeywordContainer.module.css
@@ -1,0 +1,44 @@
+.trendingTopics {
+  margin-top: 20px;
+  background-color: white;
+  border-radius: 10px;
+  padding: 20px;
+  margin-bottom: 30px;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.05);
+}
+
+.trendingTopicsHeader {
+  display: flex;
+  flex-direction: column;
+}
+
+.sectionTitle {
+  font-size: 22px;
+  font-weight: 600;
+  margin-bottom: 3px;
+  color: #333;
+}
+
+.updatedAt {
+  font-size: 14px;
+  color: #666;
+  margin-bottom: 15px;
+}
+
+.topicsList {
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+}
+
+.topicTag {
+  padding: 8px 15px;
+  background-color: var(--light-gray);
+  border-radius: 20px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+.topicTag:hover {
+  background-color: #e6e6e6;
+}
